+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v0.2 (05/05/2017)
++ start voting app
++ add a convert command, pdf2html 
++ add a html parser command, parse_html
+
 ### v0.1 (03/05/2017)
 + init repository
 + start project